+++ NEWS.md
@@ -1,5 +1,8 @@
 # dint 1.0.0.9000
 
+* added `[` method for `date_xx` objects that preserves class attribute when
+  subsetting
+
 # dint 1.0.0
 
 * First stable version